Move BoardController
